refactored code, solving yarn buld warning
