+++ app/(authenticated)/leads/layout.tsx
@@ -1,9 +1,9 @@
 'use client';
 import { useState } from 'react';
-import { ExecutiveContext } from '@/components/context/executives-context';
+import { ExecutiveContext } from '@/context/executives-context';
 import { AssignToUsers } from '@/models/global-types';
 
-export default async function LeedsLayout({ children }: { children: React.ReactNode }) {
+export default function LeedsLayout({ children }: { children: React.ReactNode }) {
   const [executivesOption, setExecutivesOption] = useState<AssignToUsers[]>([]);
 
   return (
